modify Stock

+++ StockForecastingWebApi/Models/Stock.cs
@@ -2,13 +2,13 @@
 {
     public class Stock
     {
-        public Stock(DateTime stockDate, decimal closingPrice)
+        public Stock(DateTime stockDate, float closingPrice)
         {
             StockDate=stockDate;
             ClosingPrice=closingPrice;
         }
 
         public DateTime StockDate { get; set; }
-        public decimal ClosingPrice { get; set; }
+        public float ClosingPrice { get; set; }
     }
 }
